Store edited PhraseCollection to XML
Ported definition of onClick functionality for edit button in list
selection screen to the activity itself.
Now able to save edited lists to XML and retrieve the saved lists once
the app is closed.
If no lists are currently present on the device, default lists are
loaded for the user.